feat: [EL-5832] bucket access management API

+++ api/v2/artifact.py
@@ -8,6 +8,8 @@ from botocore.exceptions import ClientError
 
 from tools import MinioClient, api_tools, auth, register_openapi
 
+from ...utils.utils import check_bucket_permission
+
 
 class ProjectAPI(api_tools.APIModeHandler):
     @register_openapi(
@@ -67,17 +69,22 @@ class ProjectAPI(api_tools.APIModeHandler):
         if not filename:
             return {'error': 'filename query parameter is required'}, 400
         decoded_filename: str = urllib.parse.unquote(filename)
-        
+
+        user = auth.current_user()
+        user_id = user.get('id') if user else None
+        if user_id and not check_bucket_permission(project_id, user_id, bucket, 'write'):
+            return {'error': 'You have read-only permission for this bucket'}, 403
+
         project = self.module.context.rpc_manager.call.project_get_or_404(project_id=project_id)
         configuration_title = request.args.get('configuration_title')
         try:
             mc = MinioClient(project, configuration_title=configuration_title)
         except AttributeError:
             return {'error': f'Error accessing s3: {configuration_title}'}, 400
-        
+
         # Delete from S3
         mc.remove_file(bucket, decoded_filename)
-        
+
         return {"message": "Deleted", "size": size(mc.get_bucket_size(bucket))}, 200
